refactor: split http result shaping
